chore(host): extract agent skill setup into agent_setup.rs
Move skill setup, plugin/hooks detection, and Claude plugin status
functions out of agent.rs into a new agent_setup.rs module.

- Keeps agent.rs focused on session scanning, live binding, and account
usage
- Makes home_path pub so agent_setup can reuse it
- Updates the one test that references the moved function
